fetch/stale-while-revalidate/stale-script.tentative.html

<!DOCTYPE html>
<!---
Tentative test against:
https://github.com/whatwg/fetch/pull/853
-->
<meta charset="utf-8">
<title>Tests Stale While Revalidate works for scripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
var last_modified;
var last_modified_count = 0;

// The script will call report via a uniquely generated ID on the subresource.
// If it is a cache hit the ID will be the same and the test will pass.
function report(mod) {
  if (!last_modified) {
    last_modified = mod;
    last_modified_count = 1;
  } else if (last_modified == mod) {
    last_modified_count++;
  }
}

async_test(t => {
  window.onload = t.step_func(() => {
    step_timeout(() => {
      var script = document.createElement("script");
      script.src = "stale-script.py";
      document.body.appendChild(script);
      script.onload = t.step_func(() => {
        var resources = performance.getEntriesByType("resource");
        var count = 0;
        for (var i=0; i < resources.length; i++) {
          if (resources[i].name.indexOf('stale-script.py') != -1) {
            count++;
          }
        }
        assert_greater_than_equal(count, 1, "performance entries");
        assert_equals(last_modified_count, 2, "last modified");

        // Ensure that we get a performance resource entry for the stale
        // while revalidate load.
        var checkResult = () => {
          var resources = performance.getEntriesByType("resource");
          var count = 0;
          for (var i=0; i < resources.length; i++) {
            if (resources[i].name.indexOf('stale-script.py') != -1) {
              count++;
            }
          }
          if (count < 2) {
            t.step_timeout(checkResult, 25);
            return;
          }

          assert_equals(count, 2, "Performance count");
          t.done();
        };

        t.step_timeout(checkResult, 25);
      });
    }, 0);
  });
}, 'Cache returns stale resource');

var script = document.createElement("script");
script.src = "stale-script.py";
document.body.appendChild(script);
</script>
</body>
